+++ demo/add.js
@@ -39,5 +39,3 @@ for (let i = 0; i < 3; i++) {
 	])
 
 }
-
-onKeyPress("space", quit)