+++ src/kaboom.ts
@@ -5691,7 +5691,7 @@ export default (gopt: KaboomOpt = {}): KaboomCtx => {
 			}
 		}
 
-		if (!assets.loaded && (gopt.loadingScreen === undefined || gopt.loadingScreen === true)) {
+		if (!assets.loaded && gopt.loadingScreen !== false) {
 		// TODO: Currently if assets are not initially loaded no updates or timers will be run, however they will run if loadingScreen is set to false. What's the desired behavior or should we make them consistent?
 			drawLoadScreen()
 		} else {