+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - (**BREAK**) rename `loadFont()` to `loadBitmapFont()`
 - added `loadFont()` to load `.ttf`, `.otf`, `.woff2` or any font supported by browser `FontFace`
 - (**BREAK**) `origin` no longer controls text alignment, use `align` option instead
+- added `quit()` to end everything
 
 ### v2000.2.6