+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - added `loadFont()` to load `.ttf`, `.otf`, `.woff2` or any font supported by browser `FontFace`
 - (**BREAK**) `origin` no longer controls text alignment, use `align` option instead
 - added `quit()` to end everything
+- shader error logs now yields the correct line number
 
 ### v2000.2.6